Export methods for working with applications to the TargetManager
